Refactor s2s-handler-babel test
