--- packages/s2s-handler-babel/src/__tests__/helpers/identifer-reverse-plugin.js
@@ -1,14 +1,0 @@
-// @flow weak
-
-export default function identifierReversePlugin() {
-  return {
-    visitor: {
-      Identifier(idPath) {
-        idPath.node.name = idPath.node.name
-          .split('')
-          .reverse()
-          .join('')
-      },
-    },
-  }
-}
