+++ packages/s2s-handler-babel/src/index.test.js
@@ -1,38 +1,28 @@
 // @flow
 import handler from '.'
-import _plugin from './__tests__/helpers/identifer-reverse-plugin'
-import fromPlugin from './__tests__/helpers/from-plugin'
+import plugin from './__tests__/helpers/plugin'
 
 const code = 'const hello = "hello"'
 
-const plugin = {
-  plugin: _plugin,
-  test: /dummy/,
+const setup = (plug = plugin) => {
+  return {
+    eventPath: '/path/to/fixture/code.js',
+    filename: 'dummy.js',
+    plugin: { test: /dummy/, plugin: plug },
+  }
 }
 
-const opts = {
-  eventPath: '/path/to/code.js',
-  filename: 'dummy.js',
-  plugin,
-}
-
-test('babel-handlerは変換後のコードを返す', () => {
-  const result = handler(code, opts)
+test('handlerは変換後のコードを返す', () => {
+  const result = handler(code, setup())
   expect(result).toMatchSnapshot()
 })
 
-test('pluginがArrayの場合', () => {
-  const result = handler(code, {
-    ...opts,
-    plugin: { ...plugin, plugin: [_plugin, { x: 1 }] },
-  })
+test('pluginがArrayの場合、変換後のCodeを返す', () => {
+  const result = handler(code, setup([plugin, { x: 1 }]))
   expect(result).toMatchSnapshot()
 })
 
-test('fromオプションを使う場合', () => {
-  const result = handler(code, {
-    ...opts,
-    plugin: { ...plugin, plugin: fromPlugin },
-  })
-  expect(result).toMatchSnapshot()
+test('codeが""の場合、""が返ること', () => {
+  const result = handler('', setup())
+  expect(result).toBe('')
 })
